Fix `RegionFile::writeHeader` chunk offset endian
This operation was not properly endian swapping, and was instead just combining the first three bytes of the integer into a singular byte via `or`.

Ideally all of these endian swap operations are centralized somewhere.

The project could also possibly be updated to utilize C++23's [std::byteswap](https://en.cppreference.com/w/cpp/numeric/byteswap) function as well to remove the need for the project to maintain it's own multi-platform byte-swap functions for each type.